Add Level() function.
