+++ logr/logger.go
@@ -10,18 +10,6 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-const (
-	EnvDevelopment = "LOG_DEVELOPMENT"
-	EnvStructured  = "LOG_STRUCTURED"
-	EnvLevel       = "LOG_LEVEL"
-)
-
-const (
-	NORMAL = 0
-	EXTRA  = 1
-	DEBUG  = 2
-)
-
 // Sink -.
 type Sink struct {
 	delegate    *log.Logger
@@ -30,27 +18,6 @@ type Sink struct {
 	development bool
 	structured  bool
 	level       int
-}
-
-// New returns a named logger.
-func New(name string, level int, kvpair ...any) logr.Logger {
-	v := os.Getenv(EnvLevel)
-	if v != "" {
-		if n, err := strconv.Atoi(v); err == nil {
-			level = n
-		}
-	}
-	return logr.New(
-		&Sink{
-			name:   name,
-			fields: fields(kvpair),
-			level:  max(NORMAL, level),
-		})
-}
-
-// WithName returns a named logger.
-func WithName(name string, kvpair ...any) logr.Logger {
-	return New(name, NORMAL, kvpair...)
 }
 
 // Init builds the delegate logger.
@@ -144,6 +111,14 @@ func (s *Sink) WithValues(kvpair ...any) logr.LogSink {
 	return &Sink{
 		name:   s.name,
 		fields: fields(kvpair),
+	}
+}
+
+func (s *Sink) WithLevel(level int) logr.LogSink {
+	return &Sink{
+		name:   s.name,
+		fields: s.fields,
+		level:  max(0, level),
 	}
 }
